mi parte de insumos integrada con la de la ro y el layout funcionando
